PB-114: Moved generateFilename in utils
