+++ src/modules/drawing/lib/export-utils.js
@@ -131,14 +131,3 @@ export function generateKmlString(projection, features = []) {
     }
     return kmlString
 }
-
-export function generateFilename(fileExtension) {
-    fileExtension = fileExtension.replace(/^\./, '')
-    const date = new Date()
-        .toISOString()
-        .split('.')[0]
-        .replaceAll('-', '')
-        .replaceAll(':', '')
-        .replace('T', '')
-    return `map.geo.admin.ch_${fileExtension.toUpperCase()}_${date}.${fileExtension.toLowerCase()}`
-}
